broke things up a bit more
