+++ src/go/handlers.go
@@ -5,20 +5,8 @@ import (
 	"encoding/json"
 	"io/ioutil"
 	"log"
-	"gopkg.in/mgo.v2"
-	"gopkg.in/mgo.v2/bson"
 	"fmt"
 )
-
-type StandardDeviationPoints struct {
-	Points []float64 `json:"points"`
-}
-
-type StandardDeviation struct {
-	StandardDeviationPoints
-	Id     bson.ObjectId `json:"id",bson:"_id,"`
-	Answer float64       `json:"answer"`
-}
 
 func HandleListStandardDeviations(w http.ResponseWriter, r *http.Request) HttpError {
 	listerr := HttpOK()
@@ -70,35 +58,4 @@ func HandlePostStandardDeviation(w http.ResponseWriter, r *http.Request) HttpErr
 	}
 
 	return posterr
-}
-
-func list() ([]StandardDeviation, error) {
-	session, err := getDb()
-	var sds []StandardDeviation = nil
-	if err == nil {
-		err = session.DB(Config.DatabaseName).C("standardDeviation").Find(nil).All(&sds)
-	}
-	defer session.Close()
-	return sds, err
-}
-func save(sd *StandardDeviation) (bool, error) {
-	success := false
-	session, err := getDb()
-	if err == nil {
-		id := bson.NewObjectId()
-		sd.Id = id
-		session.DB(Config.DatabaseName).C("standardDeviation").Insert(sd)
-		success = true
-	}
-	defer session.Close()
-	return success, nil
-}
-
-func getDb() (*mgo.Session, error) {
-	sess, err := mgo.Dial(Config.DatabaseUrl)
-	if err != nil {
-		log.Print(err)
-		return nil, err
-	}
-	return sess, nil
 }
